+++ app/src/main/java/com/diego/movies/presentation/dependency/ApplicationComponent.kt
@@ -1,6 +1,7 @@
 package com.diego.movies.presentation.dependency
 
 import com.diego.movies.App
+import com.diego.movies.presentation.dependency.application.ApplicationModule
 import dagger.Component
 import javax.inject.Singleton
